C3: usb: c3 bl33 usb host [1/1]
PD#SWPL-83818

Problem:
c3 bl33 usb host.

Solution:
c3 bl33 usb host.

Verify:
aw419

+++ drivers/usb/host/xhci-crg.c
@@ -32,8 +32,8 @@ void crg_set_mode(struct xhci_hccr *hccr, u32 mode)
 
 	if (mode == USB_DR_MODE_HOST) {
 		/* set controller host role*/
-		tmp = readl(hccr + 0x20FC) & ~0x1;
-		writel(tmp, hccr + 0x20FC);
+		tmp = readl((u64)hccr + 0x20FC) & ~0x1;
+		writel(tmp, (u64)hccr + 0x20FC);
 	}
 }
